refactor(_index.tsx, dynamic-category-icon.tsx): :truck: Move DynamicCategoryIcon component into its own file
