+++ app/routes/_index.tsx
@@ -1,17 +1,8 @@
 import type { LoaderFunction, MetaFunction } from "@remix-run/node";
 import { Link, json, useLoaderData } from "@remix-run/react";
-import type { LucideIcon } from "lucide-react";
-import {
-	AlertCircle,
-	Circle,
-	GlassWater,
-	Heart,
-	Lightbulb,
-	MapPin,
-	Salad,
-	Search,
-} from "lucide-react";
+import { Circle, MapPin } from "lucide-react";
 import { useMemo } from "react";
+import DynamicCategoryIcon from "~/components/dynamic-category-icon";
 import { Badge } from "~/components/ui/badge";
 import { Button } from "~/components/ui/button";
 import {
@@ -34,33 +25,6 @@ import {
 import VoicedeckStats from "~/components/voicedeck-stats";
 import { Report } from "~/types";
 import { fetchReports } from "../impact-reports.server";
-
-const iconComponents: { [key: string]: LucideIcon } = {
-	Hunger: Salad,
-	Thirst: GlassWater,
-	Opportunity: Lightbulb,
-	Dignity: Heart,
-};
-
-interface DynamicCategoryIconProps {
-	category: string;
-	color: string;
-	strokeWidth: string;
-	size: string;
-}
-
-const DynamicCategoryIcon: React.FC<DynamicCategoryIconProps> = ({
-	category,
-	color,
-	strokeWidth,
-	size,
-}) => {
-	const CategoryIcon = iconComponents[category];
-	if (!CategoryIcon) {
-		return <AlertCircle size={14} />; // or a placeholder component
-	}
-	return <CategoryIcon color={color} strokeWidth={strokeWidth} size={size} />;
-};
 
 export const meta: MetaFunction = () => {
 	return [
@@ -189,7 +153,6 @@ export default function Index() {
 										category={category}
 										color="#E48F85"
 										strokeWidth="1.5"
-										size="26"
 									/>
 									<p className="text-sm">{category}</p>
 								</div>
@@ -251,12 +214,11 @@ export default function Index() {
 												category={report.category}
 												color="#E48F85"
 												strokeWidth="1.5"
-												size="26"
 											/>
 											<p>{report.category}</p>
 										</Badge>
 										<Badge>
-											<MapPin color="#C14E41" strokeWidth={1} size={14} />
+											<MapPin color="#C14E41" strokeWidth={1} size={18} />
 											<p>{report.state}</p>
 										</Badge>
 									</CardContent>
